+++ test/server/inline-option.test.js
@@ -27,7 +27,7 @@ describe('inline option', () => {
     afterAll(testServer.close);
 
     it('should include inline client script in the bundle', (done) => {
-      const url = new RegExp(`client/index.js\\?http://0.0.0.0:${port}`);
+      const url = new RegExp(`client(/index.js)?\\?http://0.0.0.0:${port}`);
 
       req.get('/main.js').expect(200, url, done);
     });
@@ -54,7 +54,7 @@ describe('inline option', () => {
     afterAll(testServer.close);
 
     it('should include inline client script in the bundle', (done) => {
-      const url = new RegExp(`client/index.js\\?http://0.0.0.0:${port}`);
+      const url = new RegExp(`client(/index.js)?\\?http://0.0.0.0:${port}`);
 
       req.get('/main.js').expect(200, url, done);
     });
@@ -77,11 +77,14 @@ describe('inline option', () => {
     afterAll(testServer.close);
 
     it('should NOT include inline client script in the bundle', (done) => {
+      const url = new RegExp(`client(/index.js)?\\?http://0.0.0.0:${port}`);
+
       req
         .get('/main.js')
         .expect(200)
         .then(({ text }) => {
-          expect(text.includes(`client/index.js?http://0.0.0.0:${port}`));
+          expect(url.test(text)).toBe(false);
+
           done();
         });
     });